+++ HISTORY.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Travis Integration
+- 63 Passing
+- Parse services into policy
+
 ## [1.2.7]
 ### Added
 - Refactored test suite. See `test/_functions#run()`